Improve conditional test to check AtlasMap started and panel ready #112
previous conditional wait was checking that AtlasMap is ready and the
AtlasMap panel creation has started. This additional check is ensuring
that the panel has loaded.
Otherwise, this error might occur from time to time:
 1) Start AtlasMap Command Tests with browser type: Internal
       Test import of ADM file with running server:
     AssertionError: HTML doesn't contain url or the body tag with zero
padding: object tested must be an array, a map, an object, a set, a
string, or a weakset, but undefined given
      at testUtils.BROWSER_TYPES.forEach.testUtils.startAtlasMapInstance.then.testUtils.getWebUI.then.testUtils.startAtlasMapInstance.then.testUtils.getWebUI.then
(src/test/command.start.test.ts:223:142)
      at process._tickCallback (internal/process/next_tick.js:68:7)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>