FUSETOOLS2-45 - retain context of webview
it avoids to call a refresh when switching between tabs and getting back
to the webview. Consequently, it avoids some latency and most important
it workarounds a bug inside AtlasMap
https://github.com/atlasmap/atlasmap/issues/1455

the drawback is that the memory is still kept when the tab is hidden.
Given that for now we can open a single tab at a time, this is not a
major problem.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>