FUSETOOLS2-1280 - Avoid reloading issue in Theia
on first click after loading AtlasMap Ui in Theia, the UI is reloading.
It also reloads when the view is gaining or losing focus.

It is caused because Theia is triggering the event when gaining/losing
focus although VS Code sends it only when hiding/revealing.

A test would need to be added on Theia/Che side, do not see well how we
can test it efficiently on VS Code side.

Note: maybe we can completely remove the reload, or use a lighter
refresh mode. I think this would need to be investigated is a separated
PR as it can have a lot of impacts.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>