FUSETOOLS2-1284 - Codelens to open adm file from Camel URI in AtlasMap UI
- contains a workaround to
https://github.com/microsoft/vscode/issues/133073
- missing end-to-end test. UI tests are required because VS Code doesn't
provide API to launch codelenses.
- Supported cases:
   - First atlasmap declaration on the line. usually, there are not
several steps written on same line
  - `atlas:` (Camel 2) or `atlasmap:` (Camel 3+) scheme prefix
  - default or classpath. E.g. `atlasmap:classpath:my.adm` or
`atlasmap:my.adm`
  - The pointed _*.adm_ file is either a sibling file to the Camel Route
definition file or in `src/main/resources` at root of the workspace.
Other cases could be supported later. Given that we are not relying on
java, there isn't a concept of classpath so not possible to be complete.


Signed-off-by: Aurélien Pupier <apupier@redhat.com>